requirements.txt
anyascii==0.3.2
asgiref==3.7.2
beautifulsoup4==4.11.2
certifi==2023.5.7
charset-normalizer==3.1.0
defusedxml==0.7.1
Django==4.2.1
django-filter==22.1
django-modelcluster==6.0
django-ninja==0.22.1
django-permissionedforms==0.1
django-taggit==3.1.0
django-treebeard==4.7
djangorestframework==3.14.0
draftjs-exporter==2.1.7
et-xmlfile==1.1.0
filetype==1.2.0
html5lib==1.1
idna==3.4
l18n==2021.3
mysqlclient==2.1.1
openpyxl==3.1.2
orjson==3.9.0
Pillow==9.5.0
pydantic==1.10.8
pytz==2023.3
requests==2.31.0
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
telepath==0.3
typing_extensions==4.6.3
urllib3==2.0.2
wagtail==5.0.1
webencodings==0.5.1
Willow==1.5
